requirements.txt
flask
flask_sqlalchemy
flask_cors
werkzeug
PyJWT
python-dotenv
psycopg2-binary     # if using PostgreSQL
